scripts/requirements.txt
# Textami OOXML Parser Dependencies
# =====================================

# Core OOXML processing
python-docx>=0.8.11
lxml>=4.9.0

# JSON and data handling
jsonschema>=4.17.0

# Optional: enhanced error handling and logging
colorama>=0.4.4

# Development and testing (optional)
pytest>=7.0.0
black>=22.0.0